Complete Task 3.2: Clean up llm_tools.py - removed moved functions and added deprecation warnings
